PRD: Support to write FIR data to HOMER.
RTC: 119828
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/15972
Tested-by: Jenkins Server

+++ src/usr/diag/prdf/occ_firdata/prdfReadPnorFirData.C
@@ -71,9 +71,9 @@ class FirData
 
   private:
 
-    uint8_t  * iv_pBuf;        ///< Pointer to the PNOR data buffer
-    uint32_t   iv_maxPBufSize; ///< Maximum size of the PNOR data buffer
-    size_t     iv_pBufSize;    ///< Current size of the PNOR data buffer
+    uint8_t * iv_pBuf;        ///< Pointer to the PNOR data buffer
+    size_t    iv_maxPBufSize; ///< Maximum size of the PNOR data buffer
+    size_t    iv_pBufSize;    ///< Current size of the PNOR data buffer
 };
 
 //------------------------------------------------------------------------------
